Styles, Corvus proxying, version exposed, external URLs
- styles done
- ipc proxying for corvus
- expose version through GET parameter
- open urls marked with external hash in browser
