+++ lib/gui/pages/finish/styles/_finish.scss
@@ -26,21 +26,7 @@
 }
 
 .page-finish .button-primary {
-  min-width: $btn-min-width - 20;
-}
-
-.page-finish .button-secondary {
-  min-width: $btn-min-width - 22;
-  padding: 9px 10px;
-  background: transparent;
-  border: 2px solid $palette-theme-primary-background;
-  color: $palette-theme-primary-background;
-  font-weight: bold;
-}
-
-.page-finish .button-secondary:hover {
-  color: darken($palette-theme-primary-background, 10%);
-  border-color: darken($palette-theme-primary-background, 10%);
+  min-width: $btn-min-width;
 }
 
 .page-finish .title {
@@ -97,10 +83,8 @@
 
 .page-finish .fallback-banner {
   display: flex;
-  justify-content: center;
+  flex-direction: column;
   align-items: center;
-  align-content: flex-start;
-  flex-wrap: wrap;
   width: 100%;
   height: 300px;
   position: absolute;
@@ -108,22 +92,34 @@
   left: 0;
   background-color: $palette-theme-dark-background;
   z-index: 9001;
+  color: $palette-theme-dark-foreground;
+  font-weight: 300;
 
-  > * {
-    flex-shrink: 0;
-    padding: 10px 6px 0 6px;
+  .caption {
+    display: flex;
   }
 
-  span.caption {
-    font-size: 3rem;
-    color: $palette-theme-dark-foreground;
+  .caption-big {
+    font-size: 30px;
+  }
+
+  .caption-small {
+    font-size: 20px;
+    padding: 15px;
   }
 
   .footer-right {
     padding: 0;
+    border-bottom: 1px dashed;
   }
 
-  .caption[os-open-external] {
-    border-bottom: 2px dashed;
+  .svg-icon {
+    margin-left: 0.3em;
+    margin-right: 0.3em
+  }
+
+  .section-footer {
+    border-top: none;
+    height: 39px;
   }
 }
